# Improve LZO & LZ4 compression performance for small inputs