--- a/src/main/java/io/airlift/compress/lz4/Lz4RawCompressor.java
+++ b/src/main/java/io/airlift/compress/lz4/Lz4RawCompressor.java
@@ -13,20 +13,6 @@
  */
 package io.airlift.compress.lz4;
 
-/*
-* Licensed under the Apache License, Version 2.0 (the "License");
-* you may not use this file except in compliance with the License.
-* You may obtain a copy of the License at
-*
-*     http://www.apache.org/licenses/LICENSE-2.0
-*
-* Unless required by applicable law or agreed to in writing, software
-* distributed under the License is distributed on an "AS IS" BASIS,
-* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-* See the License for the specific language governing permissions and
-* limitations under the License.
-*/
-
 import java.util.Arrays;
 
 import static io.airlift.compress.lz4.Lz4Constants.LAST_LITERAL_SIZE;