Before:

```
  compress    airlift_lz4             silesia/dickens              6,366,945   253.2MB/s ±  8201.4kB/s ( 3.16%) (N = 30, α = 99.9%)
  compress    airlift_lz4             silesia/mozilla             26,435,667   411.5MB/s ±  8480.2kB/s ( 2.01%) (N = 30, α = 99.9%)
  compress    airlift_lz4             silesia/mr                   5,440,937   382.1MB/s ±    16.0MB/s ( 4.20%) (N = 30, α = 99.9%)
  compress    airlift_lz4             silesia/nci                  5,533,040   721.4MB/s ±    27.3MB/s ( 3.78%) (N = 30, α = 99.9%)
  compress    airlift_lz4             silesia/ooffice              4,338,918   318.9MB/s ±  9512.4kB/s ( 2.91%) (N = 30, α = 99.9%)
  compress    airlift_lz4             silesia/osdb                 5,256,666   355.4MB/s ±    10.3MB/s ( 2.89%) (N = 30, α = 99.9%)
  compress    airlift_lz4             silesia/reymont              3,181,387   263.9MB/s ±    12.2MB/s ( 4.64%) (N = 30, α = 99.9%)
  compress    airlift_lz4             silesia/samba                7,716,839   468.6MB/s ±  5566.9kB/s ( 1.16%) (N = 30, α = 99.9%)
  compress    airlift_lz4             silesia/sao                  6,790,273   305.6MB/s ±  7385.8kB/s ( 2.36%) (N = 30, α = 99.9%)
  compress    airlift_lz4             silesia/webster             19,914,645   300.8MB/s ±  6039.3kB/s ( 1.96%) (N = 30, α = 99.9%)
  compress    airlift_lz4             silesia/x-ray                8,390,195   841.3MB/s ±    23.9MB/s ( 2.85%) (N = 30, α = 99.9%)
  compress    airlift_lz4             silesia/xml                  1,227,495   542.7MB/s ±    14.5MB/s ( 2.67%) (N = 30, α = 99.9%)
  compress    airlift_lz4             artificial/a.txt                     2  2989.2kB/s ±    42.8kB/s ( 1.43%) (N = 30, α = 99.9%)   <<<<
  compress    airlift_lz4             artificial/aaa.txt                 403  9504.0MB/s ±   151.1MB/s ( 1.59%) (N = 30, α = 99.9%)   <<<<
  compress    airlift_lz4             artificial/alphabet.txt            428  6448.3MB/s ±   902.0MB/s (13.99%) (N = 30, α = 99.9%)   <<<<
  compress    airlift_lz4             artificial/random.txt          100,394  6510.8MB/s ±    71.6MB/s ( 1.10%) (N = 30, α = 99.9%)
  compress    airlift_lzo             silesia/dickens              6,314,662   202.4MB/s ±  6809.6kB/s ( 3.29%) (N = 30, α = 99.9%)
  compress    airlift_lzo             silesia/mozilla             25,511,913   363.1MB/s ±  6992.1kB/s ( 1.88%) (N = 30, α = 99.9%)
  compress    airlift_lzo             silesia/mr                   5,315,507   337.9MB/s ±  5134.6kB/s ( 1.48%) (N = 30, α = 99.9%)
  compress    airlift_lzo             silesia/nci                  5,601,480   657.1MB/s ±    15.1MB/s ( 2.30%) (N = 30, α = 99.9%)
  compress    airlift_lzo             silesia/ooffice              4,297,083   268.2MB/s ±    11.2MB/s ( 4.18%) (N = 30, α = 99.9%)
  compress    airlift_lzo             silesia/osdb                 5,283,504   290.9MB/s ±    13.2MB/s ( 4.54%) (N = 30, α = 99.9%)
  compress    airlift_lzo             silesia/reymont              2,914,188   237.1MB/s ±  3061.9kB/s ( 1.26%) (N = 30, α = 99.9%)
  compress    airlift_lzo             silesia/samba                7,516,074   385.7MB/s ±  7699.1kB/s ( 1.95%) (N = 30, α = 99.9%)
  compress    airlift_lzo             silesia/sao                  6,838,540   241.4MB/s ±  7452.2kB/s ( 3.01%) (N = 30, α = 99.9%)
  compress    airlift_lzo             silesia/webster             19,624,249   232.8MB/s ±  2861.9kB/s ( 1.20%) (N = 30, α = 99.9%)
  compress    airlift_lzo             silesia/x-ray                8,439,172  1053.4MB/s ±    14.6MB/s ( 1.38%) (N = 30, α = 99.9%)
  compress    airlift_lzo             silesia/xml                  1,183,751   477.5MB/s ±  5129.5kB/s ( 1.05%) (N = 30, α = 99.9%)
  compress    airlift_lzo             artificial/a.txt                     5  2958.3kB/s ±    43.7kB/s ( 1.48%) (N = 30, α = 99.9%)   <<<<
  compress    airlift_lzo             artificial/aaa.txt                 407  8214.3MB/s ±   126.1MB/s ( 1.53%) (N = 30, α = 99.9%)   <<<<
  compress    airlift_lzo             artificial/alphabet.txt            431  8539.4MB/s ±   137.1MB/s ( 1.61%) (N = 30, α = 99.9%)   <<<<
  compress    airlift_lzo             artificial/random.txt          100,397  6479.1MB/s ±   105.4MB/s ( 1.63%) (N = 30, α = 99.9%)
```

After:

```
  compress    airlift_lz4             silesia/dickens              6,366,945   253.4MB/s ±  6348.6kB/s ( 2.45%) (N = 30, α = 99.9%)
  compress    airlift_lz4             silesia/mozilla             26,435,667   412.3MB/s ±  6368.2kB/s ( 1.51%) (N = 30, α = 99.9%)
  compress    airlift_lz4             silesia/mr                   5,440,937   383.4MB/s ±  8135.1kB/s ( 2.07%) (N = 30, α = 99.9%)
  compress    airlift_lz4             silesia/nci                  5,533,040   723.5MB/s ±    13.6MB/s ( 1.87%) (N = 30, α = 99.9%)
  compress    airlift_lz4             silesia/ooffice              4,338,918   311.9MB/s ±    10.4MB/s ( 3.35%) (N = 30, α = 99.9%)
  compress    airlift_lz4             silesia/osdb                 5,256,666   349.0MB/s ±  6630.0kB/s ( 1.86%) (N = 30, α = 99.9%)
  compress    airlift_lz4             silesia/reymont              3,181,387   270.9MB/s ±  4325.0kB/s ( 1.56%) (N = 30, α = 99.9%)
  compress    airlift_lz4             silesia/samba                7,716,839   459.6MB/s ±  9480.0kB/s ( 2.01%) (N = 30, α = 99.9%)
  compress    airlift_lz4             silesia/sao                  6,790,273   349.0MB/s ±  4073.5kB/s ( 1.14%) (N = 30, α = 99.9%)
  compress    airlift_lz4             silesia/webster             19,914,645   287.8MB/s ±  5932.8kB/s ( 2.01%) (N = 30, α = 99.9%)
  compress    airlift_lz4             silesia/x-ray                8,390,195   899.1MB/s ±    29.6MB/s ( 3.29%) (N = 30, α = 99.9%)
  compress    airlift_lz4             silesia/xml                  1,227,495   574.5MB/s ±    13.1MB/s ( 2.28%) (N = 30, α = 99.9%)
  compress    airlift_lz4             artificial/a.txt                     2    44.9MB/s ±   940.2kB/s ( 2.05%) (N = 30, α = 99.9%)   <<<<
  compress    airlift_lz4             artificial/aaa.txt                 403    10.1GB/s ±   257.0MB/s ( 2.49%) (N = 30, α = 99.9%)   <<<<
  compress    airlift_lz4             artificial/alphabet.txt            428  8034.7MB/s ±   181.7MB/s ( 2.26%) (N = 30, α = 99.9%)   <<<<
  compress    airlift_lz4             artificial/random.txt          100,394  6421.4MB/s ±   133.1MB/s ( 2.07%) (N = 30, α = 99.9%)
  compress    airlift_lzo             silesia/dickens              6,314,662   216.1MB/s ±  6087.4kB/s ( 2.75%) (N = 30, α = 99.9%)
  compress    airlift_lzo             silesia/mozilla             25,511,913   384.4MB/s ±  9779.0kB/s ( 2.48%) (N = 30, α = 99.9%)
  compress    airlift_lzo             silesia/mr                   5,315,507   365.3MB/s ±  7847.8kB/s ( 2.10%) (N = 30, α = 99.9%)
  compress    airlift_lzo             silesia/nci                  5,601,480   686.1MB/s ±    15.9MB/s ( 2.32%) (N = 30, α = 99.9%)
  compress    airlift_lzo             silesia/ooffice              4,297,083   256.2MB/s ±  6793.8kB/s ( 2.59%) (N = 30, α = 99.9%)
  compress    airlift_lzo             silesia/osdb                 5,283,504   286.2MB/s ±  5743.7kB/s ( 1.96%) (N = 30, α = 99.9%)
  compress    airlift_lzo             silesia/reymont              2,914,188   250.4MB/s ±  6051.6kB/s ( 2.36%) (N = 30, α = 99.9%)
  compress    airlift_lzo             silesia/samba                7,516,074   407.7MB/s ±  9798.2kB/s ( 2.35%) (N = 30, α = 99.9%)
  compress    airlift_lzo             silesia/sao                  6,838,540   264.1MB/s ±  5046.5kB/s ( 1.87%) (N = 30, α = 99.9%)
  compress    airlift_lzo             silesia/webster             19,624,249   239.6MB/s ±  4530.6kB/s ( 1.85%) (N = 30, α = 99.9%)
  compress    airlift_lzo             silesia/x-ray                8,439,172  1116.7MB/s ±    32.2MB/s ( 2.88%) (N = 30, α = 99.9%)
  compress    airlift_lzo             silesia/xml                  1,183,751   506.9MB/s ±  9431.5kB/s ( 1.82%) (N = 30, α = 99.9%)
  compress    airlift_lzo             artificial/a.txt                     5    44.9MB/s ±  1083.7kB/s ( 2.36%) (N = 30, α = 99.9%)   <<<<
  compress    airlift_lzo             artificial/aaa.txt                 407  8955.3MB/s ±   231.3MB/s ( 2.58%) (N = 30, α = 99.9%)   <<<<
  compress    airlift_lzo             artificial/alphabet.txt            431  9271.0MB/s ±   239.0MB/s ( 2.58%) (N = 30, α = 99.9%)   <<<<
  compress    airlift_lzo             artificial/random.txt          100,397  6777.7MB/s ±   152.6MB/s ( 2.25%) (N = 30, α = 99.9%)
```